React 파일 위치 수정 및 gitignore 추가
