+++ Frontend/src/App.css
@@ -1,64 +1,188 @@
 .container {
   display: flex;
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 30px;
-  font-family: sans-serif;
-}
-
-.left-panel {
-  width: 40%;
-  padding-right: 20px;
-}
-
-.right-panel {
-  width: 60%;
-}
-
-.left-panel textarea {
-  width: 100%;
-  height: 350px;
-  font-family: monospace;
-  font-size: 14px;
-  padding: 10px;
-  border: 1px solid #ccc;
-  resize: none;
-}
-
-.left-panel button {
-  margin-top: 15px;
-  padding: 10px 30px;
-  font-size: 16px;
-  background-color: #ccc;
-  border: none;
-  cursor: pointer;
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.memory-area {
-  display: flex;
-  gap: 10px;
-  margin-bottom: 20px;
-}
-
-.mem-box {
-  flex: 1;
-  height: 180px;
-  border: 1px solid #aaa;
-  background-color: #fafafa;
-  padding: 10px;
+  flex-direction: row;
+  justify-content: space-between;
+  padding: 2vh 2vw;
+  gap: 5vw;
   position: relative;
 }
 
+.left-panel {
+  display: flex;
+  flex-direction: column;
+  width: 30vw;
+}
+
+.left-panel h2 {
+  text-align: center;
+  margin-bottom: 1vh;
+  font-size: 1.7vw;
+}
+
+.scrollable-container {
+  display: flex;
+  flex-direction: row;
+  height: 59.3vh;
+  border: 0.2vw solid #333;
+  background-color: #fff;
+}
+
+.line-numbers {
+  width: 1.4vw;
+  background-color: #fff;
+  text-align: right;
+  padding: 0.4vh 0.3vw 0 0;
+  font-family: 'Courier New', monospace;
+  color: #687086;
+  border-right: 2.3px solid #ccc;
+  font-size: 0.9vw;
+  line-height: 1.7vh;
+  user-select: none;
+  overflow: hidden;
+}
+
+.line-number {
+  pointer-events: none;
+  height: 1.7vh;
+}
+
+.code-input {
+  flex: 1;
+  font-family: 'Courier New', monospace;
+  font-size: 0.9vw;
+  line-height: 1.7vh;
+  padding: 0.5vh;
+  border: none;
+  outline: none;
+  resize: none;
+  overflow-y: scroll;
+  height: 100%;
+  width: 100%;
+  white-space: pre;
+  box-sizing: border-box;
+}
+
+.left-panel button {
+  margin-top: 1vh;
+  width: 30vw;
+  padding: 2vh;
+  background-color: #ddd;
+  border: none;
+  cursor: pointer;
+  font-size: 2vw;
+  border: 0.2vw solid #9e9e9e;
+  border-radius: 0.5vw;
+}
+
+.right-panel {
+  text-align: center;
+  flex-grow: 1;
+  position: relative;
+ 
+}
+
+.right-panel h2 {
+  text-align: center;
+  font-size: 1.7vw;
+  margin-bottom: 1vh;
+}
+
+.memory-container {
+  border: 0.2vw solid #333;
+  padding: 2vh;
+  min-height: 1vh;
+  width: 58vw;
+}
+
+.memory-row {
+  display: flex;
+  gap: 2vw;
+  align-items: flex-start;
+}
+
+.mem-section {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 18vw;
+  min-height: 24vh;
+
+}
+
+.mem-box {
+  border: 0.12vw solid #a0a0a0;
+  border-radius: 1vw;
+  flex-grow: 3;
+  min-height: 59vh;
+  display: flex;
+  flex-direction: column;
+  width: 16vw;
+}
+
+.mem-title {
+  padding: 1vh 2vw;
+  font-weight: bold;
+  font-size: 1.2vw;
+}
+
+.mem-content {
+  flex-grow: 1;
+  padding: 10vh;
+}
+
+.right-column {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 5vh;
+  width: 20vw;
+}
+
+.code-box {
+  min-height: 3vh;
+  height: 12vh;
+  width: 19vw;
+}
+
 .output-area {
-  background: #f0f0f0;
-  padding: 10px;
+  flex-grow: 12;
+  display: flex;
+  flex-direction: column;
+  width: 18.5vw;
 }
 
 .output-box {
-  height: 100px;
-  background-color: #aaa;
-  padding: 10px;
+  text-align: left;
+  border: 0.2vw solid #3a3a3a;
+  background-color: #444;
+  color: #fff;
+  padding: 1vh;
+  height: 30vh;
+  overflow-y: auto;
+  white-space: pre-line;
+  font-size: 0.95vw;
+}
+
+.hci-logo {
+  position: absolute;
+  bottom: -5vh;
+  right: 0vw;
+  width: 4vw;
+  height: auto;
+  opacity: 0.7;
+}
+
+.bottom-divider {
+  width: 318%;
+  height: 1px;
+  background-color: #ddd;
+  margin-top: 9vh;
+}
+
+.help-icon {
+  width: 1.1vw;
+  height: auto;
+  margin-left: 0.5vw;
+  vertical-align: middle;
+  cursor: pointer;
 }
